requirements.txt
streamlit>=1.28.0
plotly>=5.17.0
pandas>=2.0.0
groq>=0.4.0
python-dotenv>=1.0.0
pinecone-client>=2.2.0
langchain>=0.1.0
langchain-community>=0.0.1
langchain-pinecone>=0.0.1
sentence-transformers==2.2.2
huggingface-hub>=0.17.0
numpy>=1.24.0
requests>=2.31.0